Mutation

+++ netwolve/src/main/java/tom/lib/netwolve/interfaces/Provider.java
@@ -2,5 +2,5 @@ package tom.lib.netwolve.interfaces;
 
 public interface Provider {
 
-	public float provide();
+	public double provide();
 }
